+++ src/components/Login/LoginDetails.ts
@@ -1,0 +1,4 @@
+export const repoConfig = {
+  user: "MrShtrahman",
+  repo: "calculator",
+};