+++ src/redux/reducers/rootReducer.ts
@@ -1,29 +1,36 @@
+import { AuthAction, AuthType, authReducer } from "./auth/authReducer";
 import {
   CalcBasicAction,
   CalcBasicState,
-  calculatorBasicReducer
+  calculatorBasicReducer,
 } from "./calculatorBasic/calculatorBasicReducer";
 import { MemoAction, MemoState, memoReducer } from "./memo/memoReducer";
 import {
   MetadataAction,
   MetadataState,
-  calculatorMetadataReducer
+  calculatorMetadataReducer,
 } from "./calculatorMetadata/calculatorMetadataReducer";
 
 import { combineReducers } from "redux";
 
-export type RootAction = MemoAction | CalcBasicAction | MetadataAction;
+export type RootAction =
+  | MemoAction
+  | CalcBasicAction
+  | MetadataAction
+  | AuthAction;
 
 export type RootState = {
   calcBasic: CalcBasicState;
   memo: MemoState;
   calcMetadata: MetadataState;
+  auth: AuthType;
 };
 
 const rootReducer = combineReducers({
   calcBasic: calculatorBasicReducer,
   memo: memoReducer,
-  calcMetadata: calculatorMetadataReducer
+  calcMetadata: calculatorMetadataReducer,
+  auth: authReducer,
 });
 
 export default rootReducer;