+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Change Log
+
 Here all changes to the calculator are documented (**as from 05/04/2020**)
- 
+
 (The format is based on [Keep a Changelog](http://keepachangelog.com/))
- 
+
 ## Hadmama time - 05/04/2020 - 19/04/2020
-### 05/04/2020 
+
+- Adding a Login option
+- Adding CI file
 - Adding routes to the project using [**react-router-dom**](https://reacttraining.com/react-router/web/guides/quick-start) (For now, only a single route to the calculator)